+++ docker-entrypoint.sh
@@ -38,6 +38,7 @@ export NGINX_SUBFOLDER=${NGINX_SUBFOLDER:-0}
 export NGINX_SUBFOLDER_ESCAPED=$(echo ${NGINX_SUBFOLDER} | sed 's/\//\\\//g')
 export NGINX_OSB_BUCKET=${NGINX_OSB_BUCKET}
 export NGINX_OSB_RESOLVER=${NGINX_OSB_RESOLVER:-8.8.8.8}
+export HIDE_GOOGLE_GCS_HEADERS=${HIDE_GOOGLE_GCS_HEADERS:-1}
 export DRUPAL_PUBLIC_FILES_PATH=${DRUPAL_PUBLIC_FILES_PATH:-sites/default/files}
 export NGINX_CACHE_CONTROL_HEADER=${NGINX_CACHE_CONTROL_HEADER:-public,max-age=3600}
 export NGINX_GZIP_ENABLE=${NGINX_GZIP_ENABLE:-1}
@@ -67,7 +68,7 @@ if [ ! -z ${NGINX_OSB_BUCKET} ] && [ ! -f "/etc/nginx/conf.d/fragments/osb.conf"
   # We add osb.conf to fragments if Nginx is configured to use a bucket.
   # Env subst will be done later on all fragments files.
   cp /templates/fragments/osb.conf /etc/nginx/conf.d/fragments/osb.conf
-  # If we want cors, we need to add mote config to osb location.
+  # If we want cors, we need to add more config to osb location.
   if [ ${NGINX_CORS_ENABLED} == 1 ]; then
     mkdir -p /etc/nginx/conf.d/fragments/location/osb
     if [ ! -z ${NGINX_CORS_DOMAINS} ]; then
@@ -77,6 +78,12 @@ if [ ! -z ${NGINX_OSB_BUCKET} ] && [ ! -f "/etc/nginx/conf.d/fragments/osb.conf"
       print "Activating unfiltered OSB CORS"
       envsubst '${NGINX_CACHE_CONTROL_HEADER}' < /templates/fragments/location/osb/cors-unfiltered.conf > /etc/nginx/conf.d/fragments/location/osb/cors.conf
     fi
+  fi
+  # If we want to suppress google headers coming from the google storage. 
+  # We add more configuration on osb.conf file template before adding it on fragments . 
+  if [ ${HIDE_GOOGLE_GCS_HEADERS} == 1 ]; then
+    print "Hiding Google Storage headers"
+    sed  -e '/#hidegoogleheaders/r /templates/fragments/location/osb/osb-hide-google-headers.conf' -i /etc/nginx/conf.d/fragments/osb.conf;
   fi
 fi